Proposed fix for #29
This was hijacking the child ref onto the `CSSTransitionGroupChild`, this fix for #29 leaves the ref on child and lets the `TransitionGroup` add a ref to the wrapped child produced from `childFactory()`.  (assumption: `childFactory()` always wraps the child) An improvement would to detect if the child was wrapped or not and chain appropriately.